Resolving circular dependency with telemetryTypes.ts file
