+++ src/extension.ts
@@ -34,7 +34,8 @@ import { AWSStatusBar } from './shared/statusBar'
 import { AwsTelemetryOptOut } from './shared/telemetry/awsTelemetryOptOut'
 import { DefaultTelemetryService } from './shared/telemetry/defaultTelemetryService'
 import { TelemetryService } from './shared/telemetry/telemetryService'
-import { registerCommand, TelemetryNamespace } from './shared/telemetry/telemetryUtils'
+import { TelemetryNamespace } from './shared/telemetry/telemetryTypes'
+import { registerCommand } from './shared/telemetry/telemetryUtils'
 import { ExtensionDisposableFiles } from './shared/utilities/disposableFiles'
 import { PromiseSharer } from './shared/utilities/promiseUtilities'
 import { getChannelLogger } from './shared/utilities/vsCodeUtils'
