+++ src/lambda/lambdaTreeDataProvider.ts
@@ -11,7 +11,8 @@ import { AwsContextTreeCollection } from '../shared/awsContextTreeCollection'
 import { ext } from '../shared/extensionGlobals'
 import { RegionProvider } from '../shared/regions/regionProvider'
 import { ResourceFetcher } from '../shared/resourceFetcher'
-import { Datum, defaultMetricDatum, registerCommand, TelemetryNamespace } from '../shared/telemetry/telemetryUtils'
+import { Datum, TelemetryNamespace } from '../shared/telemetry/telemetryTypes'
+import { defaultMetricDatum, registerCommand } from '../shared/telemetry/telemetryUtils'
 import { AWSCommandTreeNode } from '../shared/treeview/awsCommandTreeNode'
 import { AWSTreeNodeBase } from '../shared/treeview/awsTreeNodeBase'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/awsTreeProvider'
