+++ src/shared/codelens/codeLensUtils.ts
@@ -15,7 +15,8 @@ import { SamCliProcessInvoker } from '../sam/cli/samCliInvokerUtils'
 import { SamLocalInvokeCommand } from '../sam/cli/samCliLocalInvoke'
 import { SettingsConfiguration } from '../settingsConfiguration'
 import { TelemetryService } from '../telemetry/telemetryService'
-import { Datum, defaultMetricDatum } from '../telemetry/telemetryUtils'
+import { Datum } from '../telemetry/telemetryTypes'
+import { defaultMetricDatum } from '../telemetry/telemetryUtils'
 import { toArrayAsync } from '../utilities/collectionUtils'
 import { localize } from '../utilities/vsCodeUtils'
