+++ src/shared/codelens/pythonCodeLensProvider.ts
@@ -15,7 +15,8 @@ import { fileExists, readFileAsString } from '../filesystemUtilities'
 import { LambdaHandlerCandidate } from '../lambdaHandlerSearch'
 import { getLogger } from '../logger'
 import { DefaultSamCliProcessInvoker } from '../sam/cli/samCliInvoker'
-import { Datum, registerCommand, TelemetryNamespace } from '../telemetry/telemetryUtils'
+import { Datum, TelemetryNamespace } from '../telemetry/telemetryTypes'
+import { registerCommand } from '../telemetry/telemetryUtils'
 import { getChannelLogger, getDebugPort } from '../utilities/vsCodeUtils'
 
 import { DefaultSamLocalInvokeCommand, WAIT_FOR_DEBUGGER_MESSAGES } from '../sam/cli/samCliLocalInvoke'
