+++ src/shared/codelens/typescriptCodeLensProvider.ts
@@ -12,7 +12,8 @@ import { NodejsDebugConfiguration } from '../../lambda/local/debugConfiguration'
 import { findFileInParentPaths } from '../filesystemUtilities'
 import { LambdaHandlerCandidate } from '../lambdaHandlerSearch'
 import { DefaultSamCliProcessInvoker } from '../sam/cli/samCliInvoker'
-import { Datum, registerCommand, TelemetryNamespace } from '../telemetry/telemetryUtils'
+import { Datum, TelemetryNamespace } from '../telemetry/telemetryTypes'
+import { registerCommand } from '../telemetry/telemetryUtils'
 import { TypescriptLambdaHandlerSearch } from '../typescriptLambdaHandlerSearch'
 import { getChannelLogger, getDebugPort, localize } from '../utilities/vsCodeUtils'
