+++ src/shared/telemetry/telemetryUtils.ts
@@ -7,29 +7,7 @@
 
 import * as vscode from 'vscode'
 import { ext } from '../extensionGlobals'
-import { Unit } from './clienttelemetry'
-
-export interface TelemetryName {
-    namespace: TelemetryNamespace | OldTelemetryNamespace
-    name: string
-}
-
-export interface Datum {
-    name: string
-    value: number
-    unit?: Unit
-    metadata?: Map<string, string>
-}
-
-type OldTelemetryNamespace = 'Command'
-
-export enum TelemetryNamespace {
-    Aws = 'aws',
-    Cloudformation = 'cloudformation',
-    Lambda = 'lambda',
-    Project = 'project',
-    Session = 'session'
-}
+import { Datum, TelemetryName } from './telemetryTypes'
 
 export function defaultMetricDatum(name: string): Datum {
     return {
